FEAT: Scanning functionality and checking of the scanned vouchers

+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/repository/voucher/dto/Voucher.kt
@@ -7,7 +7,7 @@ data class Voucher(
     var qrCode: String = "",
     var vendorId: String = "",
     var productIds: Array<Long> = arrayOf(),
-    var price: Long = 0,
+    var price: Double = 0.0,
     var currency: String = "",
     var value: Long = 0,
     var booklet: String = "",
